Fix: Resolve remaining analyzer errors in Flame components and services
- Add missing await super.onLoad() in Flame component onLoad() methods
  * skill_visual_effect.dart: SkillVisualEffect, CriticalBurst, BurstParticle
  * damage_number.dart: DamageNumber, ComboCounter
  * Ensures game reference is available during component initialization

- Fix Flame component constructor calls
  * DamageNumber: Pass position/size/anchor to super() call
  * Remove redundant field initialization in onLoad()

- Fix audioplayers API usage
  * bgm_service.dart: Remove getVolume() call (not available)
  * Use tracked _volume instead

- Fix RevenueCat offerings iteration
  * monetization_service.dart: Use offerings.all.values for Map iteration

- Fix performance service memory monitoring
  * Remove invalid Service.getVM() call
  * Use placeholder value for now

- Fix type errors in services and viewmodels
  * skill_system_service.dart: Remove .toDouble() calls, change 'def' to 'defense'
  * battle_viewmodel.dart: Convert currentMana to int with toInt()

- Fix CustomButton parameter names
  * skill_build_screen.dart: Change text→label, onTap→onPressed

- Remove unused imports
  * skill_build_screen.dart: Remove app_config import
  * rendering_optimization.dart: Remove flame/game and flutter/foundation imports

+++ lib/viewmodels/battle_viewmodel.dart
@@ -291,7 +291,7 @@ class BattleViewModel extends StateNotifier<BattleState> {
 
     if (selfParticipant != null && state.playerResources != null) {
       final updated = state.playerResources!.copyWith(
-        currentMana: selfParticipant.resources?.currentMana ?? 100.0,
+        currentMana: (selfParticipant.resources?.currentMana ?? 100).toInt(),
         gold: selfParticipant.resources?.gold ?? 0,
         ownedItemIds: selfParticipant.resources?.ownedItemIds ?? [],
       );
